feat: correct feature gating, including text
