+++ Cargo.toml
@@ -57,23 +57,24 @@ repository.workspace = true
 bevy = { workspace = true }
 # Ensure this version stays up to date with the README
 vello = "0.4.0"
-vello_svg = "0.6.0"
-velato = "0.5.0"
 thiserror = "2.0.11"
-once_cell = "1.19.0"
-bytemuck = { version = "1.21.0", features = ["derive"] }
+vello_svg = { version = "0.6.0", optional = true }
+velato = { version = "0.5.0", optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-parley = { version = "0.3.0", features = ["system"] }
+parley = { version = "0.3.0", optional = true, features = ["system"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-parley = { version = "0.3.0", default-features = false, features = ["std"] }
+parley = { version = "0.3.0", optional = true, default-features = false, features = [
+  "std",
+] }
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.50"
 
 [features]
-default = ["default_font"]
-svg = []
-lottie = []
-default_font = ["bevy/default_font"]
+default = []
+svg = ["vello_svg"]
+lottie = ["velato"]
+text = ["parley"]
+default_font = ["bevy/default_font", "text"]
