+++ examples/text/Cargo.toml
@@ -8,5 +8,5 @@ publish = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
-bevy_vello = { path = "../../", features = ["default_font"] }
+bevy_vello = { path = "../../", features = ["text", "default_font"] }
 bevy = { workspace = true, features = ["default_font", "bevy_gizmos"] }
