+++ src/integrations/text/context.rs
