+++ src/integrations/text/font.rs
@@ -1,16 +1,14 @@
-use std::borrow::Cow;
-
-use crate::text::context::{LOCAL_LAYOUT_CONTEXT, get_global_font_context};
-
 use super::{
     VelloTextAnchor,
     context::LOCAL_FONT_CONTEXT,
     vello_text::{VelloFontAxes, VelloTextSection},
 };
+use crate::integrations::text::context::{LOCAL_LAYOUT_CONTEXT, get_global_font_context};
 use bevy::{prelude::*, reflect::TypePath, render::render_asset::RenderAsset};
 use parley::{
     FontSettings, FontStyle, FontVariation, PositionedLayoutItem, RangedBuilder, StyleProperty,
 };
+use std::borrow::Cow;
 use vello::{
     Scene,
     kurbo::Affine,
