+++ src/integrations/text/font_loader.rs
@@ -1,5 +1,5 @@
 use super::{context::get_global_font_context, font::VelloFont};
-use crate::{integrations::VectorLoaderError, text::context::LOCAL_FONT_CONTEXT};
+use crate::{integrations::VectorLoaderError, integrations::text::context::LOCAL_FONT_CONTEXT};
 use bevy::{
     asset::{AssetLoader, LoadContext, io::Reader},
     log::warn,
