+++ src/integrations/text/render.rs
@@ -1,0 +1,134 @@
+use super::{VelloFont, VelloTextAnchor, VelloTextSection};
+use crate::render::{SkipEncoding, VelloEntityCountData, VelloView, prepare::PreparedAffine};
+use bevy::{
+    prelude::*,
+    render::{
+        Extract,
+        camera::ExtractedCamera,
+        sync_world::TemporaryRenderEntity,
+        view::{ExtractedView, RenderLayers},
+    },
+};
+use vello::kurbo::Affine;
+
+#[derive(Component, Clone)]
+pub struct ExtractedVelloText {
+    pub text: VelloTextSection,
+    pub text_anchor: VelloTextAnchor,
+    pub transform: GlobalTransform,
+}
+
+pub fn extract_text(
+    mut commands: Commands,
+    query_views: Query<
+        (&ExtractedCamera, Option<&RenderLayers>),
+        (With<Camera2d>, With<VelloView>),
+    >,
+    query_scenes: Extract<
+        Query<
+            (
+                &VelloTextSection,
+                &VelloTextAnchor,
+                &GlobalTransform,
+                &ViewVisibility,
+                &InheritedVisibility,
+                Option<&RenderLayers>,
+            ),
+            Without<SkipEncoding>,
+        >,
+    >,
+    fonts: Extract<Res<Assets<VelloFont>>>,
+    mut frame_data: ResMut<VelloEntityCountData>,
+) {
+    let mut n_texts = 0;
+
+    // Sort cameras by rendering order
+    let mut views: Vec<_> = query_views.iter().collect();
+    views.sort_unstable_by_key(|(camera, _)| camera.order);
+
+    for (text, text_anchor, transform, view_visibility, inherited_visibility, render_layers) in
+        query_scenes.iter()
+    {
+        // Skip if visibility conditions are not met
+        if !view_visibility.get() || !inherited_visibility.get() {
+            continue;
+        }
+        // Skip if font isn't loaded.
+        let Some(_font) = fonts.get(text.style.font.id()) else {
+            continue;
+        };
+
+        // Check if any camera renders this asset
+        let asset_render_layers = render_layers.unwrap_or_default();
+        if views.iter().any(|(_, camera_layers)| {
+            asset_render_layers.intersects(camera_layers.unwrap_or_default())
+        }) {
+            commands
+                .spawn(ExtractedVelloText {
+                    text: text.clone(),
+                    text_anchor: *text_anchor,
+                    transform: *transform,
+                })
+                .insert(TemporaryRenderEntity);
+            n_texts += 1;
+        }
+    }
+
+    frame_data.n_texts = n_texts;
+}
+
+pub fn prepare_text_affines(
+    mut commands: Commands,
+    views: Query<(&ExtractedCamera, &ExtractedView), (With<Camera2d>, With<VelloView>)>,
+    render_entities: Query<(Entity, &ExtractedVelloText)>,
+) {
+    for (camera, view) in views.iter() {
+        let size_pixels: UVec2 = camera.physical_viewport_size.unwrap();
+        let (pixels_x, pixels_y) = (size_pixels.x as f32, size_pixels.y as f32);
+
+        for (entity, render_entity) in render_entities.iter() {
+            let ndc_to_pixels_matrix = Mat4::from_cols_array_2d(&[
+                [pixels_x / 2.0, 0.0, 0.0, pixels_x / 2.0],
+                [0.0, pixels_y / 2.0, 0.0, pixels_y / 2.0],
+                [0.0, 0.0, 1.0, 0.0],
+                [0.0, 0.0, 0.0, 1.0],
+            ])
+            .transpose();
+
+            let world_transform = render_entity.transform;
+
+            let mut model_matrix = world_transform.compute_matrix();
+            model_matrix.w_axis.y *= -1.0;
+
+            let (projection_mat, view_mat) = {
+                let mut view_mat = view.world_from_view.compute_matrix();
+                view_mat.w_axis.y *= -1.0;
+
+                (view.clip_from_view, view_mat)
+            };
+
+            let view_proj_matrix = projection_mat * view_mat.inverse();
+            let vello_matrix = ndc_to_pixels_matrix * view_proj_matrix;
+
+            let raw_transform = vello_matrix * model_matrix;
+
+            let transform: [f32; 16] = raw_transform.to_cols_array();
+
+            // | a c e |
+            // | b d f |
+            // | 0 0 1 |
+            let transform: [f64; 6] = [
+                transform[0] as f64,  // a
+                -transform[1] as f64, // b
+                -transform[4] as f64, // c
+                transform[5] as f64,  // d
+                transform[12] as f64, // e
+                transform[13] as f64, // f
+            ];
+
+            commands
+                .entity(entity)
+                .insert(PreparedAffine(Affine::new(transform)));
+        }
+    }
+}
