+++ src/plugin.rs
@@ -1,7 +1,6 @@
 use crate::{
-    VelloFont, VelloRenderSettings,
+    VelloRenderSettings,
     render::{VelloCanvasSettings, VelloRenderPlugin},
-    text::VelloFontLoader,
 };
 use bevy::{prelude::*, render::view::RenderLayers};
 use vello::AaConfig;
@@ -40,24 +39,12 @@ impl Plugin for VelloPlugin {
                 use_cpu: self.use_cpu,
                 antialiasing: self.antialiasing,
             },
-        })
-        .init_asset::<VelloFont>()
-        .init_asset_loader::<VelloFontLoader>();
+        });
         #[cfg(feature = "svg")]
         app.add_plugins(crate::integrations::svg::SvgIntegrationPlugin);
         #[cfg(feature = "lottie")]
         app.add_plugins(crate::integrations::lottie::LottieIntegrationPlugin);
-        #[cfg(feature = "default_font")]
-        {
-            let mut fonts = app
-                .world_mut()
-                .get_resource_mut::<Assets<VelloFont>>()
-                .unwrap();
-
-            fonts.insert(
-                Handle::default().id(),
-                crate::text::load_into_font_context(bevy::text::DEFAULT_FONT_DATA.to_vec()),
-            );
-        }
+        #[cfg(feature = "text")]
+        app.add_plugins(crate::integrations::text::VelloTextIntegrationPlugin);
     }
 }
