+++ src/render/extract.rs
@@ -77,72 +77,6 @@ pub fn extract_scenes(
     frame_data.n_scenes = n_scenes;
 }
 
-#[derive(Component, Clone)]
-pub struct ExtractedVelloText {
-    pub text: VelloTextSection,
-    pub text_anchor: VelloTextAnchor,
-    pub transform: GlobalTransform,
-}
-
-pub fn extract_text(
-    mut commands: Commands,
-    query_views: Query<
-        (&ExtractedCamera, Option<&RenderLayers>),
-        (With<Camera2d>, With<VelloView>),
-    >,
-    query_scenes: Extract<
-        Query<
-            (
-                &VelloTextSection,
-                &VelloTextAnchor,
-                &GlobalTransform,
-                &ViewVisibility,
-                &InheritedVisibility,
-                Option<&RenderLayers>,
-            ),
-            Without<SkipEncoding>,
-        >,
-    >,
-    fonts: Extract<Res<Assets<VelloFont>>>,
-    mut frame_data: ResMut<VelloEntityCountData>,
-) {
-    let mut n_texts = 0;
-
-    // Sort cameras by rendering order
-    let mut views: Vec<_> = query_views.iter().collect();
-    views.sort_unstable_by_key(|(camera, _)| camera.order);
-
-    for (text, text_anchor, transform, view_visibility, inherited_visibility, render_layers) in
-        query_scenes.iter()
-    {
-        // Skip if visibility conditions are not met
-        if !view_visibility.get() || !inherited_visibility.get() {
-            continue;
-        }
-        // Skip if font isn't loaded.
-        let Some(_font) = fonts.get(text.style.font.id()) else {
-            continue;
-        };
-
-        // Check if any camera renders this asset
-        let asset_render_layers = render_layers.unwrap_or_default();
-        if views.iter().any(|(_, camera_layers)| {
-            asset_render_layers.intersects(camera_layers.unwrap_or_default())
-        }) {
-            commands
-                .spawn(ExtractedVelloText {
-                    text: text.clone(),
-                    text_anchor: *text_anchor,
-                    transform: *transform,
-                })
-                .insert(TemporaryRenderEntity);
-            n_texts += 1;
-        }
-    }
-
-    frame_data.n_texts = n_texts;
-}
-
 /// Synchronize the entity count data back to the main world.
 pub fn sync_entity_count(render_data: Res<VelloEntityCountData>, mut world: ResMut<MainWorld>) {
     let mut main_world_data = world.get_resource_mut::<VelloEntityCountData>().unwrap();
