--- src/text/mod.rs
@@ -1,10 +1,0 @@
-//! Components and text logic.
-
-mod context;
-mod font;
-mod font_loader;
-mod vello_text;
-
-pub use font::VelloFont;
-pub(crate) use font_loader::{VelloFontLoader, load_into_font_context};
-pub use vello_text::{VelloTextAlign, VelloTextAnchor, VelloTextSection, VelloTextStyle};
